+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The current file describes significant changes in the code that may affect the u
 #### PlatformBundle
 * Added `\Oro\Bundle\PlatformBundle\Validator\Constraints\ValidEmbeddable` that allows to apply `Valid` constraint with explicit validation groups specified in `embeddedGroups` option.
 
+#### DataGridBundle
+* Added postponed delete entities logic to `\Oro\Bundle\DataGridBundle\Extension\MassAction\DeleteMassActionHandler`.
 
 ### Deprecated